% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/amatrix.R
\name{amatrix}
\alias{amatrix}
\title{amatrix Function}
\usage{
amatrix(ped)
}
\arguments{
\item{ped}{a data frame of pedigree with three columns: ID, SIRE, DAM}
}
\description{
This function allows you to calculate A matrix, the numerical relationship matrix.
}
\examples{
amatrix()
}
\keyword{amatrix}
