% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biosim-package.R
\docType{package}
\name{biosim-package}
\alias{biosim}
\alias{biosim-package}
\title{biosim: a bunch of functions in quantitative genetics.}
\description{
a collection of functions in quantitative genetics.
}
\author{
\strong{Maintainer}: Yongjun Li \email{Yongjun.li@agriculture.vic.gov.au}

}
\keyword{internal}
